backend/requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
flask==2.2.5
flask-cors==3.0.10
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
fairlearn>=0.7.0
aif360>=0.4.0
nltk>=3.8.1
pdfminer.six>=20221105
python-dotenv>=0.21.1
gunicorn>=20.1.0
